Fixed more compiler warnings.
Still more warnings about 64 bit sizes going into normal ints.

+++ src/framework/fs/pathutils.cpp
@@ -84,7 +84,7 @@ std::list<char*>* get_path_parts(const char* path)
 
 void sanitize_path_slashes(const char* original, char* sanitized)
 {
-    int len = strlen(original);
+    int len = (int)strlen(original);
 
     for(int i = 0; i < len; i++)
     {
